+++ src/error.rs
@@ -4,7 +4,7 @@ use std::fmt;
 use std::num::{IntErrorKind, ParseIntError};
 
 /// An error when parsing from a string.
-#[derive(Debug)]
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub enum ParseError {
     /// An integer failed to parse.
     ParseInt {
@@ -44,6 +44,24 @@ impl fmt::Display for ParseError {
                 key,
                 number,
             } => write!(f, "Unknown {key}: {number}"),
+        }
+    }
+}
+
+/// An error when appending to a set.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum InsertError {
+    /// The set is full.
+    Full,
+    /// The item is already in the set.
+    Duplicate,
+}
+
+impl fmt::Display for InsertError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            InsertError::Full => write!(f, "Set is full."),
+            InsertError::Duplicate => write!(f, "Item already exists."),
         }
     }
 }